Fixes for false splits on the final bosses
Occasionally when the level restarts after a death, the boss HP values will be set to 0 for a frame, resulting in a false split. Checks to make sure the player had control on the previous update to avoid this (current update can still potentially cause an issue, so previous is used instead).